Initial changes for Left 4 Dead 2.
* Added/removed various functions from the following interfaces: IVEngineServer, IServerGameDLL, IServerGameClients, IFilesystem, and IConVar
* Added new functions to ConVar for setting Color values
* Updated SendProp structure with unknown member
* Recompiled tier1 static library for Windows and Linux
* Fixed up some tier0 exports in public/tier0/dbg.h
* Rebuilt tier0 import libary for Windows
* Replaced tier0_i486.so and vstdlib_i486.so with the newly renamed versions: tier0_linux.so and vstdlib_linux.so

--HG--
rename : lib/linux/tier0_i486.so => lib/linux/tier0_linux.so
rename : lib/linux/vstdlib_i486.so => lib/linux/vstdlib_linux.so

+++ public/igameevents.h
@@ -75,13 +75,13 @@ public:
 	// Data access
 	virtual bool  GetBool( const char *keyName = NULL, bool defaultValue = false ) = 0;
 	virtual int   GetInt( const char *keyName = NULL, int defaultValue = 0 ) = 0;
-	virtual unsigned long long GetUint64 ( const char *keyName = NULL, unsigned long long defaultValue = 0) = 0;
+	virtual uint64 GetUint64 ( const char *keyName = NULL, unsigned long long defaultValue = 0) = 0;
 	virtual float GetFloat( const char *keyName = NULL, float defaultValue = 0.0f ) = 0;
 	virtual const char *GetString( const char *keyName = NULL, const char *defaultValue = "" ) = 0;
 
 	virtual void SetBool( const char *keyName, bool value ) = 0;
 	virtual void SetInt( const char *keyName, int value ) = 0;
-	virtual void SetUint64( const char *keyName, unsigned long long value ) = 0;
+	virtual void SetUint64( const char *keyName, uint64 value ) = 0;
 	virtual void SetFloat( const char *keyName, float value ) = 0;
 	virtual void SetString( const char *keyName, const char *value ) = 0;
 };
